requirements.txt
streamlit>=1.24.0
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
scikit-learn>=0.24.2
xlrd>=2.0.1
statsmodels>=0.13.0
prophet==1.1.2
cmdstanpy==1.0.4
tensorflow-cpu==2.10.0
